+++ lib/isodoc/gb/gbwordconvert.rb
@@ -1,10 +1,6 @@
-require "isodoc"
-require_relative "gbconvert"
-require_relative "gbcleanup"
-require "gb_agencies"
-require_relative "metadata"
+require_relative "gbbaseconvert"
 require_relative "gbwordrender"
-require "fileutils"
+require "isodoc"
 
 module IsoDoc
   module Gb
@@ -44,49 +40,6 @@ module IsoDoc
         }
       end 
 
-      def extract_fonts(options)
-        b = options[:bodyfont] || "Arial"
-        h = options[:headerfont] || "Arial"
-        m = options[:monospacefont] || "Courier"
-        t = options[:titlefont] || "Arial"
-        "$bodyfont: #{b};\n$headerfont: #{h};\n$monospacefont: #{m};\n$titlefont: #{t};\n"
-      end   
-
-      def metadata_init(lang, script, labels)
-        unless ["en", "zh"].include? lang
-          lang = "zh"
-          script = "Hans"
-        end
-        @meta = Metadata.new(lang, script, labels)
-        @meta.set(:standardclassimg, @standardclassimg)
-        @common.meta = @meta
-      end
-
-      def cleanup(docxml)
-        @cleanup = Cleanup.new(@script, @deprecated_lbl)
-        super
-        @cleanup.cleanup(docxml)
-        docxml
-      end
-
-      def example_cleanup(docxml)
-        super
-        @cleanup.example_cleanup(docxml)
-      end
-
-      def i18n_init(lang, script)
-        super
-        y = if lang == "en"
-              YAML.load_file(File.join(File.dirname(__FILE__), "i18n-en.yaml"))
-            elsif lang == "zh" && script == "Hans"
-              YAML.load_file(File.join(File.dirname(__FILE__),
-                                       "i18n-zh-Hans.yaml"))
-            else
-              YAML.load_file(File.join(File.dirname(__FILE__), "i18n-zh-Hans.yaml"))
-            end
-        @labels = @labels.merge(y)
-      end
-
       ENDLINE = <<~END.freeze
       <v:line id="_x0000_s1026"
  alt="" style='position:absolute;left:0;text-align:left;z-index:251662848;
@@ -107,9 +60,7 @@ module IsoDoc
         docxml
       end
 
-      def omit_docid_prefix(prefix)
-        super || prefix == "Chinese Standard"
-      end
+      include BaseConvert
     end
   end
 end
